+++ public_html/admin/install/include/install.lib.php
@@ -6,7 +6,7 @@
 // |                                                                          |
 // | glFusion installation script.                                            |
 // +--------------------------------------------------------------------------+
-// | Copyright (C) 2008-2016 by the following authors:                        |
+// | Copyright (C) 2008-2017 by the following authors:                        |
 // |                                                                          |
 // | Mark R. Evans          mark AT glfusion DOT org                          |
 // | Eric Warren            eric AT glfusion DOT org                          |
@@ -1432,6 +1432,7 @@ function INST_doDatabaseUpgrades($current_fusion_version, $use_innodb = false)
             $c->add('enable_404_logging',1,'select',7,3,0,20,TRUE);
             $c->add('debug_oauth',0,'select',7,3,0,30,TRUE);
             $c->add('debug_html_filter',0,'select',7,3,0,40,TRUE);
+            $c->add('dbback_exclude','', 'text',0,0,NULL,1,TRUE,'dbadmin_internal');
 
             $current_fusion_version = '1.6.5';